refactored menu, fixed overdraft/minbalance logic
