+++ src/bankapp/BankAppLauncher.java
@@ -2,7 +2,7 @@ package bankapp;
 
 public class BankAppLauncher {
     public static void main(String[] args) {
-        Menu menu = new Menu();
-        menu.startMenu();
+        MainMenu mainMenu = new MainMenu();
+        mainMenu.startMenu();
     }
 }
